relocate guice annotations to new package
