+++ megabus/src/main/java/com/bazaarvoice/megabus/MegabusBootWorkflowManager.java
@@ -5,9 +5,11 @@ import com.bazaarvoice.emodb.common.dropwizard.guice.SelfHostAndPort;
 import com.bazaarvoice.emodb.common.dropwizard.leader.LeaderServiceTask;
 import com.bazaarvoice.emodb.common.dropwizard.lifecycle.LifeCycleRegistry;
 import com.bazaarvoice.emodb.kafka.Topic;
+import com.bazaarvoice.megabus.guice.MegabusTopic;
+import com.bazaarvoice.megabus.guice.MegabusZookeeper;
 import com.bazaarvoice.megabus.refproducer.MegabusRefProducerManager;
-import com.bazaarvoice.megabus.resolver.guice.MegabusRefResolverService;
-import com.bazaarvoice.megabus.resolver.guice.MissingRefDelayService;
+import com.bazaarvoice.megabus.guice.MegabusRefResolverService;
+import com.bazaarvoice.megabus.guice.MissingRefDelayService;
 import com.google.common.net.HostAndPort;
 import com.google.common.util.concurrent.Service;
 import com.google.inject.Inject;
