+++ megabus/src/main/java/com/bazaarvoice/megabus/MegabusModule.java
@@ -4,13 +4,16 @@ import com.bazaarvoice.emodb.common.dropwizard.log.DefaultRateLimitedLogFactory;
 import com.bazaarvoice.emodb.common.dropwizard.log.RateLimitedLogFactory;
 import com.bazaarvoice.emodb.kafka.KafkaCluster;
 import com.bazaarvoice.emodb.kafka.Topic;
+import com.bazaarvoice.megabus.guice.MegabusRefTopic;
+import com.bazaarvoice.megabus.guice.MegabusTopic;
+import com.bazaarvoice.megabus.guice.MissingRefTopic;
+import com.bazaarvoice.megabus.guice.RetryRefTopic;
 import com.bazaarvoice.megabus.refproducer.MegabusRefProducerConfiguration;
 import com.bazaarvoice.megabus.refproducer.MegabusRefProducerManager;
 import com.bazaarvoice.megabus.refproducer.MegabusRefSubscriptionMonitorManager;
 import com.bazaarvoice.megabus.refproducer.NumRefPartitions;
-import com.bazaarvoice.megabus.resolver.DocumentResolverManager;
-import com.bazaarvoice.megabus.resolver.guice.MegabusRefResolverService;
-import com.bazaarvoice.megabus.resolver.guice.MissingRefDelayService;
+import com.bazaarvoice.megabus.guice.MegabusRefResolverService;
+import com.bazaarvoice.megabus.guice.MissingRefDelayService;
 import com.bazaarvoice.megabus.resolver.ResilientMegabusRefResolver;
 import com.bazaarvoice.megabus.resolver.ResilientMissingRefDelayProcessor;
 import com.google.common.collect.ImmutableMap;
