+++ megabus/src/main/java/com/bazaarvoice/megabus/guice/MissingRefTopic.java
@@ -1,4 +1,4 @@
-package com.bazaarvoice.megabus.resolver.guice;
+package com.bazaarvoice.megabus.guice;
 
 import com.google.inject.BindingAnnotation;
 import java.lang.annotation.Retention;
@@ -10,5 +10,5 @@ import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
 @BindingAnnotation
 @Target({ FIELD, PARAMETER, METHOD }) @Retention(RUNTIME)
-public @interface MegabusRefResolverService {
+public @interface MissingRefTopic {
 }
