+++ megabus/src/main/java/com/bazaarvoice/megabus/refproducer/MegabusRefProducerManager.java
@@ -19,8 +19,8 @@ import com.bazaarvoice.emodb.kafka.KafkaCluster;
 import com.bazaarvoice.emodb.kafka.Topic;
 import com.bazaarvoice.emodb.sor.condition.Conditions;
 import com.bazaarvoice.megabus.MegabusApplicationId;
-import com.bazaarvoice.megabus.MegabusRefTopic;
-import com.bazaarvoice.megabus.MegabusZookeeper;
+import com.bazaarvoice.megabus.guice.MegabusRefTopic;
+import com.bazaarvoice.megabus.guice.MegabusZookeeper;
 import com.codahale.metrics.MetricRegistry;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.net.HostAndPort;
