+++ megabus/src/main/java/com/bazaarvoice/megabus/resolver/MissingRefDelayProcessor.java
@@ -7,13 +7,12 @@ import com.bazaarvoice.emodb.kafka.Topic;
 import com.bazaarvoice.emodb.sor.core.DataProvider;
 import com.bazaarvoice.megabus.MegabusApplicationId;
 import com.bazaarvoice.megabus.MegabusRef;
-import com.bazaarvoice.megabus.MissingRefTopic;
-import com.bazaarvoice.megabus.RetryRefTopic;
+import com.bazaarvoice.megabus.guice.MissingRefTopic;
+import com.bazaarvoice.megabus.guice.RetryRefTopic;
 import com.bazaarvoice.megabus.service.KafkaStreamsService;
 import com.codahale.metrics.MetricRegistry;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.google.common.net.HostAndPort;
-import com.google.inject.Inject;
 import java.time.Clock;
 import java.time.Instant;
 import java.util.List;
