+++ megabus/src/main/java/com/bazaarvoice/megabus/resolver/ResilientMegabusRefResolver.java
@@ -5,10 +5,10 @@ import com.bazaarvoice.emodb.kafka.KafkaCluster;
 import com.bazaarvoice.emodb.kafka.Topic;
 import com.bazaarvoice.emodb.sor.core.DataProvider;
 import com.bazaarvoice.megabus.MegabusApplicationId;
-import com.bazaarvoice.megabus.MegabusRefTopic;
-import com.bazaarvoice.megabus.MegabusTopic;
-import com.bazaarvoice.megabus.MissingRefTopic;
-import com.bazaarvoice.megabus.RetryRefTopic;
+import com.bazaarvoice.megabus.guice.MegabusRefTopic;
+import com.bazaarvoice.megabus.guice.MegabusTopic;
+import com.bazaarvoice.megabus.guice.MissingRefTopic;
+import com.bazaarvoice.megabus.guice.RetryRefTopic;
 import com.bazaarvoice.megabus.service.ResilientService;
 import com.codahale.metrics.MetricRegistry;
 import com.google.common.net.HostAndPort;
