+++ megabus/src/main/java/com/bazaarvoice/megabus/resolver/ResilientMissingRefDelayProcessor.java
@@ -5,9 +5,8 @@ import com.bazaarvoice.emodb.kafka.KafkaCluster;
 import com.bazaarvoice.emodb.kafka.Topic;
 import com.bazaarvoice.emodb.sor.core.DataProvider;
 import com.bazaarvoice.megabus.MegabusApplicationId;
-import com.bazaarvoice.megabus.MissingRefTopic;
-import com.bazaarvoice.megabus.RetryRefTopic;
-import com.bazaarvoice.megabus.resolver.MissingRefDelayProcessor;
+import com.bazaarvoice.megabus.guice.MissingRefTopic;
+import com.bazaarvoice.megabus.guice.RetryRefTopic;
 import com.bazaarvoice.megabus.service.ResilientService;
 import com.codahale.metrics.MetricRegistry;
 import com.google.common.net.HostAndPort;
