+++ web/src/main/java/com/bazaarvoice/emodb/web/EmoModule.java
@@ -106,7 +106,6 @@ import com.bazaarvoice.emodb.web.resources.databus.SubjectDatabus;
 import com.bazaarvoice.emodb.web.resources.databus.SubjectDatabusClientFactory;
 import com.bazaarvoice.emodb.web.scanner.ScanUploadModule;
 import com.bazaarvoice.emodb.web.scanner.ScannerZooKeeper;
-import com.bazaarvoice.emodb.web.scanner.config.ScannerConfiguration;
 import com.bazaarvoice.emodb.web.settings.DatabusDefaultJoinFilterConditionAdminTask;
 import com.bazaarvoice.emodb.web.settings.Setting;
 import com.bazaarvoice.emodb.web.settings.SettingsModule;
@@ -127,10 +126,9 @@ import com.bazaarvoice.emodb.web.throttling.ZkAdHocThrottleSerializer;
 import com.bazaarvoice.emodb.web.throttling.ZkDataStoreUpdateThrottleSerializer;
 import com.bazaarvoice.emodb.web.util.ZKNamespaces;
 import com.bazaarvoice.megabus.MegabusApplicationId;
-import com.bazaarvoice.megabus.MegabusBootDAO;
 import com.bazaarvoice.megabus.MegabusConfiguration;
 import com.bazaarvoice.megabus.MegabusModule;
-import com.bazaarvoice.megabus.MegabusZookeeper;
+import com.bazaarvoice.megabus.guice.MegabusZookeeper;
 import com.bazaarvoice.ostrich.HostDiscovery;
 import com.bazaarvoice.ostrich.MultiThreadedServiceFactory;
 import com.bazaarvoice.ostrich.ServiceEndPoint;
