+++ web/src/main/java/com/bazaarvoice/emodb/web/megabus/MegabusStashModule.java
@@ -26,9 +26,8 @@ import com.bazaarvoice.emodb.web.scanner.writer.ScanWriterGenerator;
 import com.bazaarvoice.emodb.web.util.ZKNamespaces;
 import com.bazaarvoice.megabus.MegabusBootConfiguration;
 import com.bazaarvoice.megabus.MegabusBootDAO;
-import com.bazaarvoice.megabus.MegabusZookeeper;
+import com.bazaarvoice.megabus.guice.MegabusZookeeper;
 import com.google.common.base.Optional;
-import com.google.inject.Key;
 import com.google.inject.PrivateModule;
 import com.google.inject.Provides;
 import com.google.inject.Singleton;
